update:fixbug

+++ src/main/java/com/ggj/article/module/business/bean/Media.java
@@ -28,7 +28,7 @@ public class Media extends BaseEntity {
 
     private String publishSpeed;
 
-    private Byte baiduSeo;
+    private int baiduSeo;
 
     private String exampleUrl;
